Replace cGZPersistDBSegmentMultiPackedFiles
The built-in cGZPersistDBSegmentMultiPackedFiles class that the game
uses when loading DAT files from sub-directories can be a bottleneck
when scanning directories and looking up the loaded TGI values.
Our replacement class is much faster in both of those cases.

There are currently limitations on the file names that can be loaded,
they must be US-ASCII.
This is likely due to a difference in the UTF-8<->UTF-16 string
conversion between the Windows API calls that we use and the custom
code that SC4 uses.