Remove the increased buffer size patches
This has been replaced by a patch to the ReadWithCount method that
skips using the buffer when it is empty and the requested read size is
larger than the buffer size.
Reverting this change should also save a significant amount of memory
when a large number of files are open.