Skip over DAT files that fail to load
This ensures that the other plugins will still be read.
The DLL will write the path of any files that fail to load to the log.

A MultiPackedFile will be considered to have loaded successfully if at
least one DAT file was loaded from the folders that were searched.
This is similar logic to what the base game uses, and will allow it to
skip MultiPackedFiles that are for folders with no DAT files in them.