Simplify the map code that inserts the TGI values
Because we are using our own cIGZPersistResourceKeyList implementation,
we do not need to use the cIGZPersistResourceKeyList::EnumKeys method.

Grabbing a const reference to the internal collection that the
PersistResourceKeyList class uses to store the data and enumerating
over it directly may allow the compiler to produce better code.
The PersistResourceKeyList class has also been marked as final.