Fix the crash when reading uncompressed records
The return type of cIGZPersistDBSegment::ReadRecord should be uint32_t
not bool.
ReadRecord returns the record size using both the return value and an
output parameter.

The crash was likely caused by the game using the truncated return
value as the total record size, and failing to read all of the data.