Use NtQueryDirectoryFile for directory searching
NtQueryDirectoryFile is a lower level API than FindFirstFile, it should
be faster wen dealing with folders containing large numbers of files.
It also allows us to reuse the same buffer when recursively searching
each directory.

If NtQueryDirectoryFile is not available the FindFirstFile API will be
used as a fallback, but this case should never occur as the plugin does
not support the DOS-based versions of Windows.