Fix loading DBPF files without an extension
The Bosham Church lot by mintoes uses a .SC4Lot file without a file
extension.

This works in the original game code because of the way that Maxis
wrote the code to load .SC4* files (.SC4Desc, .SC4Model, etc.), the
game will gather all files from the plugins folder and sub-folders
regardless of file extension and then try to load anything that
doesn't have a .DAT extension as a .SC4* file.

Our fix is to treat any file without an extension as a .SC4* file.
If the file is not a valid DBPF file, the game will fail to load it
and the plugin will report that as an error.