Fix a bug where too little data could be read
This is not an issue 99% of the time, as the OS will usually return all
of the requested data in a single ReadFile call.

The original game code reads data in a loop until the OS returns all of
the requested data, an error occurs, or the end of the file is reached.

We previously used a similar read loop, but the code was removed due to
an infinite loop bug when the end of the file was reached unexpectedly.
This reverts that commit (8d9c522adf416dd1ea4b2d517236d5fd0fe43433) and
fixes the infinite loop bug.