Revert 'Use NtQueryDirectoryFile for directory searching'
Some users were having issues with plugins failing to load that was
caused by a bug in the NtQueryDirectoryFile code.

The plugin will now use FindFirstFile/FindNextFile for directory
searching, same as version 0.9 and earlier.
FindFirstFile/FindNextFile is slower than NtQueryDirectoryFile for
directories with a massive number of files and folders
(tens of thousands), but few if any users will have a plugins folder
large enough for that to be a significant performance issue.

This reverts commit 1a95825342c95b845fac78f8f263b398a2b1754a.