Replace SC4's SetupResources method with a multi-threaded version
The 'user data' folder (typically Documents/SimCity 4/Plugins) is
scanned for .SC4* and .DAT plugins on background threads, while the
main thread loads the other DBPF files.
This assumes that users don't place a large number of plugins in the
SC4 installation directory plugins folders, <Language>, Sku_data, and
Plugins.

The background scanning is started before loading other DBPF files, and
the code will wait for the background scanning to finish when it needs
to add that data to the resource manager.

The .SC4* and .DAT plugins can be read into their respective
MultiPackedFile class without any multi-threading concerns.
The only GZCOM method the code calls is the thread-safe GetClassObject,
and the cGZDBSegmentPackedFile instance it creates is only accessed by
one thread at a time.